arm64: dts: ti: k3-j784s4-main: Add ACSPCIE1 node
The ACSPCIE1 module on TI's J784S4 SoC is capable of driving the reference
clock required by the PCIe Endpoint device. It is an alternative to on-
board and external reference clock generators.
Add the device-tree node for the same.

Upstream-Status: Submitted [https://lore.kernel.org/all/20250513152155.1590689-1-parth105105@gmail.com/]
Signed-off-by: Parth Pancholi <parth.pancholi@toradex.com>