arm64: dts: ti: k3-am69-aquila: Add active cooling support
Add the necessary SoM temperature trip points and fan configuration for
the supported carrier boards. Below is a description of the active
cooling behavior and how the implementation was tested.

At room temperature, with the system idle, the fan speed duty cycle is
40%, with no audible noise from the fan, and the T-junction is
maintained around 55 ºC. Running GPU and CPU stress tests to raise the
SoC temperature, the fan speed increases to a 70% duty cycle when the
T-junction reaches approximately 70 ºC. At this point, with a 70% duty
cycle, the temperature is maintained at ~70 ºC. The fan runs at 100%
duty cycle when the T-junction exceeds 85 ºC, but it was not possible to
raise the SoC temperature above ~80 ºC at room temperature with active
cooling enabled.

Testing of the various active trip points was performed using the
thermal emulation mode provided by Linux (Kconfig:
CONFIG_THERMAL_EMULATION). The thermal governor used during
implementation and testing was the Step governor.

Worth mentioning that around the trip points temperature, the fan speed
can keep switching between the configured speed for each trip point
until the trip point crossing is complete. This behavior is related to
the implementation of thermal governors in the kernel, where the
governors are not handling hysteresis correctly. The issue seems to be
fixed after v6.10, see [1] and [2].

[1] https://lore.kernel.org/lkml/12346243.O9o76ZdvQC@kreacher/
[2] https://lore.kernel.org/lkml/13515747.uLZWGnKmhe@kreacher/

Upstream-Status: Pending
Signed-off-by: João Paulo Gonçalves <joao.goncalves@toradex.com>