Revert "drm/tidss: Adjust the pclk based on the HW capabilities"
The LVDS output clock on Verdin AM62 is incorrect when using the native
LVDS interface.
The expected pixel clock is 71.1 MHz, but due to this commit, the actual
output clock is only 24.2 MHz. Reverting it restores the correct pixel
clock.
The reverted commit is part of a patch series that is still under
discussion and will be reworked in the future.
The issue addressed by this revert was already reported on the mailing
list in response to the v2 series [1], even though v3 of the patch
series had already been sent [2].

[1] https://lore.kernel.org/all/20250402-cdns-dsi-impro-v2-3-4a093eaa5e27@ideasonboard.com/
[2] https://lore.kernel.org/all/20250414-cdns-dsi-impro-v3-3-4e52551d4f07@ideasonboard.com/

This reverts commit 986e5529d4cc11b73d33bb40b1799a0dc46bad88.

Upstream-Status: Inappropriate
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>