drm/panel: st7701: Add Winstar wf40eswaa6mnn0 panel support
The Winstar wf40eswaa6mnn0 panel is a square 4.0" TFT LCD with a
resolution of 480x480 pixels.

This panel is driven by the Sitronix ST7701 controller and uses a MIPI
DSI interface. The settings are based on the panel's datasheet and the
init sequence provided by Winstar.

It was tested on a Verdin iMX8MP from Toradex with a Carrier Board
providing a MIPI DSI interface.

Signed-off-by: Stefan Eichenberger <eichest@gmail.com>
Reviewed-by: Neil Armstrong <neil.armstrong@linaro.org>
Signed-off-by: Neil Armstrong <neil.armstrong@linaro.org>
Link: https://lore.kernel.org/r/20250606114644.105371-2-eichest@gmail.com
(cherry picked from commit d04f6367d39918461d0335d30b860d38668d4b54)

[ Backport: Renamed macros to match symbol names in kernel 6.6.
  Macros with `ST7701_CMD2_` prefix to `DSI_CMD2_`.
  `ST7701_WRITE` macros to `ST7701_DSI`. ]

Signed-off-by: Parth Pancholi <parth.pancholi@toradex.com>