drm/panel: st7703: Add Pila display panel support
Add support for the Pila display panel with a resolution of 720x1440
pixels, driven by the Sitronix ST7703 controller over a MIPI DSI
interface. This is the first working patch and serves as a starting
point; timings need further confirmation. Tested on an AM62 Verdin
board from Toradex with a MIPI DSI interface.

Note: This patch includes hacks such as non-continuous clock mode and
initialization in the prepare phase, which require reworking to
provide a decent upstreamable solution.

Signed-off-by: Parth Pancholi <parth.pancholi@toradex.com>