arm64: dts: ti: k3-am69-aquila: Update regulators for v1.1 HW revision
Update PMIC and regulator configuration to address the differences
between Aquila AM69 v1.0 vs v1.1.

On V1.1
 - PMIC BUCK5 is used for 3.3V
 - PMIC LDO2 is used for 2.5V ethernet
 - Wi-Fi power needs to be enabled with GPIO0_57, this change is applied
   to both v1.0 and v1.1 for simplicity, since on v1.0 this GPIO is not
   used therefore there is no side effect on doing it.

For simplicity remove 1.0V ethernet regulator that is not used for
anything and it's different between v1.0 and v1.1.

Wi-Fi regulator is currently not properly described, we would need a
vpcie-supply as a child of the PCIe root port, but this is not
implemented. In the meantime mark this regulator as always on, please
note that this might now work as expected.

Upstream-Status: Pending
Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>